Fixed packet handling and showing realmlist.

+++ Common/Network/Packets/PacketReader.cs
@@ -7,7 +7,7 @@ namespace Common.Network.Packets
 {
     public class PacketReader : BinaryReader
     {
-        public uint Opcode { get; set; }
+        public int Opcode { get; set; }
         public ushort Size { get; set; }
 
         public PacketReader(byte[] data, bool worldPacket = true) : base(new MemoryStream(data))
@@ -15,7 +15,7 @@ namespace Common.Network.Packets
             if (worldPacket)
             {
                 Size = (ushort)(this.ReadUInt16() - 4);
-                Opcode = this.ReadUInt32();
+                Opcode = (int)this.ReadUInt32();
 
                 if (Opcode != (uint)ClientMessage.TransferInitiate)
                     this.ReadUInt16();
